refactor(frontend): [#30] Folder structure changed , sidebar width issue fixed
