+++ src/types/sheet.ts
@@ -8,6 +8,7 @@ export interface SheetProps {
   onClose: () => void;
   position?: 'top' | 'right' | 'bottom' | 'left';
   onOpenChange?: (open: boolean) => void;
+  width?: number;
 }
 
 export interface UseSheetProps {
